0.368
linux下写日志文件尚存问题

+++ src/convert.cpp
@@ -28,7 +28,7 @@ namespace Matrix
 
     std::string Convert::Int2Str(int value)
     {
-        return std::to_string(value);
+        return std::to_string((long long int)value);
     }
 
     int Convert::Str2Int(std::string value)
